+++ fabric-fluid-v1/src/main/java/net/fabricmc/fabric/impl/fluid/BucketItemAccessor.java
@@ -1,0 +1,7 @@
+package net.fabricmc.fabric.impl.fluid;
+
+import net.minecraft.fluid.Fluid;
+
+public interface BucketItemAccessor {
+	Fluid getFluid();
+}
