--- fabric-fluid-v1/src/main/java/net/fabricmc/fabric/impl/fluids/BucketItemAccessor.java
@@ -1,7 +1,0 @@
-package net.fabricmc.fabric.impl.fluids;
-
-import net.minecraft.fluid.Fluid;
-
-public interface BucketItemAccessor {
-	Fluid getFluid();
-}
